refactor: move dock shell API to workflow module
